+++ src/lib.rs
@@ -453,3 +453,4 @@ pub use conversion::quick_collection;
 
 pub type JsonValue = serde_json::Value;
 pub type JsonObject = serde_json::Map<String, JsonValue>;
+pub type MalformedGeoJsonError = serde_json::Error;